refactor: flatten src/legacy/plugins/ — discord and http promoted to legacy/ children
Plugin system was removed in #747; the only things left under
src/legacy/plugins/ were discord/ and http/, neither of which are
plugins anymore (both instantiated directly from app.ts). Promotes them
to first-class legacy/ children and deletes the now-empty plugins/
directory.

Pure relocation. Internal up-paths re-anchored, version.js import in
http/status.ts adjusted from ../../ to ../, gateway imports inside
legacy/{discord,http} simplified from ../../legacy/gateway/ to
../gateway/. CLAUDE.md updated.
