+++ src/legacy/discord/config.test.ts
