+++ src/legacy/discord/config.ts
