+++ src/legacy/discord/isotopes.plugin.json
