+++ src/legacy/discord/message-metadata.ts
