+++ src/legacy/http/chat.test.ts
@@ -4,10 +4,10 @@ import { randomUUID } from "node:crypto";
 import { describe, it, expect, beforeEach, afterEach } from "vitest";
 import http from "node:http";
 import { ApiServer } from "./server.js";
-import { CronScheduler } from "../../../automation/cron-job.js";
-import { SessionStoreManager } from "../../../agent/runners/pi/session-store.js";
-import { AgentRuntime } from "../../../agent/runtime.js";
-import { createMockSessionStore } from "../../../test-helpers.js";
+import { CronScheduler } from "../../automation/cron-job.js";
+import { SessionStoreManager } from "../../agent/runners/pi/session-store.js";
+import { AgentRuntime } from "../../agent/runtime.js";
+import { createMockSessionStore } from "../../test-helpers.js";
 
 const MOCK_AGENT_ID = "mock";
