+++ src/legacy/http/cron.ts
@@ -2,7 +2,7 @@
 
 import { addRoute } from "./routes.js";
 import { sendJson, sendError, handleRouteError } from "./middleware.js";
-import type { CronJobInput } from "../../../automation/cron-job.js";
+import type { CronJobInput } from "../../automation/cron-job.js";
 
 // ---------------------------------------------------------------------------
 // GET /api/cron — list cron jobs
