+++ src/legacy/http/middleware.ts
@@ -2,7 +2,7 @@
 // CORS, JSON body parsing, error handling, and request logging.
 
 import type { IncomingMessage, ServerResponse } from "node:http";
-import { createLogger } from "../../../logging/logger.js";
+import { createLogger } from "../../logging/logger.js";
 
 const log = createLogger("api");
