+++ src/legacy/http/server.test.ts
@@ -3,7 +3,7 @@
 import { describe, it, expect, beforeEach, afterEach } from "vitest";
 import http from "node:http";
 import { ApiServer } from "./server.js";
-import { CronScheduler } from "../../../automation/cron-job.js";
+import { CronScheduler } from "../../automation/cron-job.js";
 
 // ---------------------------------------------------------------------------
 // Helpers
