+++ src/legacy/http/sessions.ts
@@ -9,13 +9,13 @@
 
 import { addRoute } from "./routes.js";
 import { sendJson, sendError } from "./middleware.js";
-import { createLogger } from "../../../logging/logger.js";
+import { createLogger } from "../../logging/logger.js";
 import { randomUUID } from "node:crypto";
-import { runAgent } from "../../../agent/runtime-adapter.js";
-import { userMessage } from "../../../agent/runners/pi/messages.js";
-import { resolveAgentWorkspacePath } from "../../../paths.js";
-import type { DefaultSessionStore } from "../../../agent/runners/pi/session-store.js";
-import type { Session } from "../../../sessions/types.js";
+import { runAgent } from "../../agent/runtime-adapter.js";
+import { userMessage } from "../../agent/runners/pi/messages.js";
+import { resolveAgentWorkspacePath } from "../../paths.js";
+import type { DefaultSessionStore } from "../../agent/runners/pi/session-store.js";
+import type { Session } from "../../sessions/types.js";
 
 const log = createLogger("api:sessions");
