+++ src/legacy/http/static.ts
