+++ src/legacy/http/status.ts
@@ -2,7 +2,7 @@
 
 import { addRoute } from "./routes.js";
 import { sendJson } from "./middleware.js";
-import { VERSION } from "../../version.js";
+import { VERSION } from "../version.js";
 
 addRoute("GET", "/api/status", (_req, res, deps) => {
   const cronJobCount = deps.cronScheduler.listJobs().length;
